+++ crates/teloxide-core/src/types/successful_payment.rs
@@ -1,6 +1,15 @@
+use derive_more::derive::From;
 use serde::{Deserialize, Serialize};
 
 use crate::types::OrderInfo;
+
+/// A unique identifier of the Telegram transaction.
+#[derive(Clone, Debug, derive_more::Display)]
+#[derive(PartialEq, Eq, Hash)]
+#[derive(Serialize, Deserialize, From)]
+#[serde(transparent)]
+#[from(&'static str, String)]
+pub struct TelegramTransactionId(pub String);
 
 /// This object contains basic information about a successful payment.
 ///
@@ -35,8 +44,28 @@ pub struct SuccessfulPayment {
     pub order_info: OrderInfo,
 
     /// Telegram payment identifier.
-    pub telegram_payment_charge_id: String,
+    pub telegram_payment_charge_id: TelegramTransactionId,
 
     /// Provider payment identifier.
     pub provider_payment_charge_id: String,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn deser() {
+        let telegram_payment_charge_id =
+            S { telegram_payment_charge_id: TelegramTransactionId("id".to_owned()) };
+        let json = r#"{"telegram_payment_charge_id":"id"}"#;
+
+        #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+        struct S {
+            telegram_payment_charge_id: TelegramTransactionId,
+        }
+
+        assert_eq!(serde_json::to_string(&telegram_payment_charge_id).unwrap(), json);
+        assert_eq!(telegram_payment_charge_id, serde_json::from_str(json).unwrap());
+    }
 }