+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for TBA 7.6 ([#1356](https://github.com/teloxide/teloxide/pull/1356))
   - Add `filter_paid_media` to `MessageFilterExt` trait
 
+- Support for TBA 7.7 ([#1357](https://github.com/teloxide/teloxide/pull/1357)) [**BC**]
+
 ### Changed
 
 - New id types ([#1153](https://github.com/teloxide/teloxide/pull/1153)) [**BC**]